review fixes-- remove evalQueue, fox console evaluation, etc
